# Add optimized paths to CRON parsing