* seal CronExpression
* add .NET 6 target
* spanify APIs

## Quartz.Benchmark.CronExpressionBenchmark

| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |CreateExpressionsAndCalculateFireTimeAfter|5.674 μs|0.0091 μs|4.73 KB|
| **New** |	| **5.355 μs (-6%)** | **0.0156 μs** | **3.57 KB (-25%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|3.849 μs|0.0090 μs|2.66 KB|
| **New** |	| **3.488 μs (-9%)** | **0.0298 μs** | **1.71 KB (-36%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|5.465 μs|0.0155 μs|5.45 KB|
| **New** |	| **4.994 μs (-9%)** | **0.0498 μs** | **3.5 KB (-36%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|6.106 μs|0.0110 μs|3.24 KB|
| **New** |	| **5.606 μs (-8%)** | **0.0357 μs** | **2.09 KB (-35%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|6.631 μs|0.0086 μs|4.16 KB|
| **New** |	| **6.116 μs (-8%)** | **0.0470 μs** | **2.8 KB (-33%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|6.588 μs|0.0137 μs|4.16 KB|
| **New** |	| **6.078 μs (-8%)** | **0.0593 μs** | **2.8 KB (-33%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|6.321 μs|0.0152 μs|3.55 KB|
| **New** |	| **5.979 μs (-5%)** | **0.0323 μs** | **2.32 KB (-35%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|6.716 μs|0.0141 μs|3.84 KB|
| **New** |	| **6.183 μs (-8%)** | **0.0630 μs** | **2.64 KB (-31%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|6.984 μs|0.0153 μs|4.13 KB|
| **New** |	| **6.427 μs (-8%)** | **0.0505 μs** | **2.87 KB (-31%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|5.798 μs|0.0098 μs|2.46 KB|
| **New** |	| **5.369 μs (-7%)** | **0.0088 μs** | **1.27 KB (-48%)** |

/cc @jafin